SFO Passenger Volume EDA and Prediction
This is a simple KNIME workflow for EDA and time series prediction. The GroupBy node is used to quickly produce some statistics. I like to use the R Plot node as it gives me a lot of flexibility to customize the plot.
Start the Rserve per KNIME's documentation before running this workflow.